refactor(agents): split subagent registry
